+++ logicrunner/builtin/proxy/costcenter/costcenter.go
@@ -59,10 +59,8 @@ func GetPrototype() insolar.Reference {
 }
 
 // New is constructor
-func New(commissionWallet insolar.Reference, currentTariff insolar.Reference) *ContractConstructorHolder {
-	var args [2]interface{}
-	args[0] = commissionWallet
-	args[1] = currentTariff
+func New() *ContractConstructorHolder {
+	var args [0]interface{}
 
 	var argsSerialized []byte
 	err := common.CurrentProxyCtx.Serialize(args, &argsSerialized)
@@ -137,98 +135,15 @@ func (r *CostCenter) GetCode() (insolar.Reference, error) {
 	return r.Code, nil
 }
 
-// SetTariffs is proxy generated method
-func (r *CostCenter) SetTariffs(tariffs []insolar.Reference) error {
+// CalcFee is proxy generated method
+func (r *CostCenter) CalcFee(amountStr string) (string, error) {
 	var args [1]interface{}
-	args[0] = tariffs
-
-	var argsSerialized []byte
-
-	ret := [1]interface{}{}
-	var ret0 *foundation.Error
-	ret[0] = &ret0
-
-	err := common.CurrentProxyCtx.Serialize(args, &argsSerialized)
-	if err != nil {
-		return err
-	}
-
-	res, err := common.CurrentProxyCtx.RouteCall(r.Reference, true, false, false, "SetTariffs", argsSerialized, *PrototypeReference)
-	if err != nil {
-		return err
-	}
-
-	err = common.CurrentProxyCtx.Deserialize(res, &ret)
-	if err != nil {
-		return err
-	}
-
-	if ret0 != nil {
-		return ret0
-	}
-	return nil
-}
-
-// SetTariffsNoWait is proxy generated method
-func (r *CostCenter) SetTariffsNoWait(tariffs []insolar.Reference) error {
-	var args [1]interface{}
-	args[0] = tariffs
-
-	var argsSerialized []byte
-
-	err := common.CurrentProxyCtx.Serialize(args, &argsSerialized)
-	if err != nil {
-		return err
-	}
-
-	_, err = common.CurrentProxyCtx.RouteCall(r.Reference, false, false, false, "SetTariffs", argsSerialized, *PrototypeReference)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// SetTariffsAsImmutable is proxy generated method
-func (r *CostCenter) SetTariffsAsImmutable(tariffs []insolar.Reference) error {
-	var args [1]interface{}
-	args[0] = tariffs
-
-	var argsSerialized []byte
-
-	ret := [1]interface{}{}
-	var ret0 *foundation.Error
-	ret[0] = &ret0
-
-	err := common.CurrentProxyCtx.Serialize(args, &argsSerialized)
-	if err != nil {
-		return err
-	}
-
-	res, err := common.CurrentProxyCtx.RouteCall(r.Reference, true, true, false, "SetTariffs", argsSerialized, *PrototypeReference)
-	if err != nil {
-		return err
-	}
-
-	err = common.CurrentProxyCtx.Deserialize(res, &ret)
-	if err != nil {
-		return err
-	}
-
-	if ret0 != nil {
-		return ret0
-	}
-	return nil
-}
-
-// GetTariffs is proxy generated method
-func (r *CostCenter) GetTariffs() ([]insolar.Reference, error) {
-	var args [0]interface{}
+	args[0] = amountStr
 
 	var argsSerialized []byte
 
 	ret := [2]interface{}{}
-	var ret0 []insolar.Reference
+	var ret0 string
 	ret[0] = &ret0
 	var ret1 *foundation.Error
 	ret[1] = &ret1
@@ -238,7 +153,7 @@ func (r *CostCenter) GetTariffs() ([]insolar.Reference, error) {
 		return ret0, err
 	}
 
-	res, err := common.CurrentProxyCtx.RouteCall(r.Reference, true, false, false, "GetTariffs", argsSerialized, *PrototypeReference)
+	res, err := common.CurrentProxyCtx.RouteCall(r.Reference, true, false, false, "CalcFee", argsSerialized, *PrototypeReference)
 	if err != nil {
 		return ret0, err
 	}
@@ -254,9 +169,10 @@ func (r *CostCenter) GetTariffs() ([]insolar.Reference, error) {
 	return ret0, nil
 }
 
-// GetTariffsNoWait is proxy generated method
-func (r *CostCenter) GetTariffsNoWait() error {
-	var args [0]interface{}
+// CalcFeeNoWait is proxy generated method
+func (r *CostCenter) CalcFeeNoWait(amountStr string) error {
+	var args [1]interface{}
+	args[0] = amountStr
 
 	var argsSerialized []byte
 
@@ -265,7 +181,7 @@ func (r *CostCenter) GetTariffsNoWait() error {
 		return err
 	}
 
-	_, err = common.CurrentProxyCtx.RouteCall(r.Reference, false, false, false, "GetTariffs", argsSerialized, *PrototypeReference)
+	_, err = common.CurrentProxyCtx.RouteCall(r.Reference, false, false, false, "CalcFee", argsSerialized, *PrototypeReference)
 	if err != nil {
 		return err
 	}
@@ -273,14 +189,15 @@ func (r *CostCenter) GetTariffsNoWait() error {
 	return nil
 }
 
-// GetTariffsAsImmutable is proxy generated method
-func (r *CostCenter) GetTariffsAsImmutable() ([]insolar.Reference, error) {
-	var args [0]interface{}
+// CalcFeeAsImmutable is proxy generated method
+func (r *CostCenter) CalcFeeAsImmutable(amountStr string) (string, error) {
+	var args [1]interface{}
+	args[0] = amountStr
 
 	var argsSerialized []byte
 
 	ret := [2]interface{}{}
-	var ret0 []insolar.Reference
+	var ret0 string
 	ret[0] = &ret0
 	var ret1 *foundation.Error
 	ret[1] = &ret1
@@ -290,176 +207,7 @@ func (r *CostCenter) GetTariffsAsImmutable() ([]insolar.Reference, error) {
 		return ret0, err
 	}
 
-	res, err := common.CurrentProxyCtx.RouteCall(r.Reference, true, true, false, "GetTariffs", argsSerialized, *PrototypeReference)
-	if err != nil {
-		return ret0, err
-	}
-
-	err = common.CurrentProxyCtx.Deserialize(res, &ret)
-	if err != nil {
-		return ret0, err
-	}
-
-	if ret1 != nil {
-		return ret0, ret1
-	}
-	return ret0, nil
-}
-
-// SetCurrentTariff is proxy generated method
-func (r *CostCenter) SetCurrentTariff(currentTariff insolar.Reference) error {
-	var args [1]interface{}
-	args[0] = currentTariff
-
-	var argsSerialized []byte
-
-	ret := [1]interface{}{}
-	var ret0 *foundation.Error
-	ret[0] = &ret0
-
-	err := common.CurrentProxyCtx.Serialize(args, &argsSerialized)
-	if err != nil {
-		return err
-	}
-
-	res, err := common.CurrentProxyCtx.RouteCall(r.Reference, true, false, false, "SetCurrentTariff", argsSerialized, *PrototypeReference)
-	if err != nil {
-		return err
-	}
-
-	err = common.CurrentProxyCtx.Deserialize(res, &ret)
-	if err != nil {
-		return err
-	}
-
-	if ret0 != nil {
-		return ret0
-	}
-	return nil
-}
-
-// SetCurrentTariffNoWait is proxy generated method
-func (r *CostCenter) SetCurrentTariffNoWait(currentTariff insolar.Reference) error {
-	var args [1]interface{}
-	args[0] = currentTariff
-
-	var argsSerialized []byte
-
-	err := common.CurrentProxyCtx.Serialize(args, &argsSerialized)
-	if err != nil {
-		return err
-	}
-
-	_, err = common.CurrentProxyCtx.RouteCall(r.Reference, false, false, false, "SetCurrentTariff", argsSerialized, *PrototypeReference)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// SetCurrentTariffAsImmutable is proxy generated method
-func (r *CostCenter) SetCurrentTariffAsImmutable(currentTariff insolar.Reference) error {
-	var args [1]interface{}
-	args[0] = currentTariff
-
-	var argsSerialized []byte
-
-	ret := [1]interface{}{}
-	var ret0 *foundation.Error
-	ret[0] = &ret0
-
-	err := common.CurrentProxyCtx.Serialize(args, &argsSerialized)
-	if err != nil {
-		return err
-	}
-
-	res, err := common.CurrentProxyCtx.RouteCall(r.Reference, true, true, false, "SetCurrentTariff", argsSerialized, *PrototypeReference)
-	if err != nil {
-		return err
-	}
-
-	err = common.CurrentProxyCtx.Deserialize(res, &ret)
-	if err != nil {
-		return err
-	}
-
-	if ret0 != nil {
-		return ret0
-	}
-	return nil
-}
-
-// GetCurrentTariff is proxy generated method
-func (r *CostCenter) GetCurrentTariff() (insolar.Reference, error) {
-	var args [0]interface{}
-
-	var argsSerialized []byte
-
-	ret := [2]interface{}{}
-	var ret0 insolar.Reference
-	ret[0] = &ret0
-	var ret1 *foundation.Error
-	ret[1] = &ret1
-
-	err := common.CurrentProxyCtx.Serialize(args, &argsSerialized)
-	if err != nil {
-		return ret0, err
-	}
-
-	res, err := common.CurrentProxyCtx.RouteCall(r.Reference, true, false, false, "GetCurrentTariff", argsSerialized, *PrototypeReference)
-	if err != nil {
-		return ret0, err
-	}
-
-	err = common.CurrentProxyCtx.Deserialize(res, &ret)
-	if err != nil {
-		return ret0, err
-	}
-
-	if ret1 != nil {
-		return ret0, ret1
-	}
-	return ret0, nil
-}
-
-// GetCurrentTariffNoWait is proxy generated method
-func (r *CostCenter) GetCurrentTariffNoWait() error {
-	var args [0]interface{}
-
-	var argsSerialized []byte
-
-	err := common.CurrentProxyCtx.Serialize(args, &argsSerialized)
-	if err != nil {
-		return err
-	}
-
-	_, err = common.CurrentProxyCtx.RouteCall(r.Reference, false, false, false, "GetCurrentTariff", argsSerialized, *PrototypeReference)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// GetCurrentTariffAsImmutable is proxy generated method
-func (r *CostCenter) GetCurrentTariffAsImmutable() (insolar.Reference, error) {
-	var args [0]interface{}
-
-	var argsSerialized []byte
-
-	ret := [2]interface{}{}
-	var ret0 insolar.Reference
-	ret[0] = &ret0
-	var ret1 *foundation.Error
-	ret[1] = &ret1
-
-	err := common.CurrentProxyCtx.Serialize(args, &argsSerialized)
-	if err != nil {
-		return ret0, err
-	}
-
-	res, err := common.CurrentProxyCtx.RouteCall(r.Reference, true, true, false, "GetCurrentTariff", argsSerialized, *PrototypeReference)
+	res, err := common.CurrentProxyCtx.RouteCall(r.Reference, true, true, false, "CalcFee", argsSerialized, *PrototypeReference)
 	if err != nil {
 		return ret0, err
 	}
